+++ src/fides/api/app_setup.py
@@ -83,8 +83,9 @@ def create_fides_app(
         # Starlette bug causing this to fail mypy
         fastapi_app.add_exception_handler(FunctionalityNotConfigured, handler)  # type: ignore
     fastapi_app.add_middleware(SlowAPIMiddleware)
-    fastapi_app.add_middleware(GZipMiddleware, minimum_size=1000, compresslevel=5)  # minimum_size is in bytes
-
+    fastapi_app.add_middleware(
+        GZipMiddleware, minimum_size=1000, compresslevel=5
+    )  # minimum_size is in bytes
 
     for router in routers:
         fastapi_app.include_router(router)