+++ CHANGELOG.md
@@ -17,7 +17,9 @@ The types of changes are:
 
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.43.1...main)
 
+### Added
 
+- Added Gzip Middleware for responses [#5225](https://github.com/ethyca/fides/pull/5225)
 
 ## [2.43.1](https://github.com/ethyca/fides/compare/2.43.0...2.43.1)